+++ edge-hub/core/src/Microsoft.Azure.Devices.Edge.Hub.MqttBrokerAdapter/handlers/PolicyUpdateHandler.cs
@@ -43,21 +43,31 @@ namespace Microsoft.Azure.Devices.Edge.Hub.MqttBrokerAdapter
             var config = await configSource.GetConfig();
             await config.ForEachAsync(async config =>
             {
-                var update = ConfigToPolicyUpdate(config);
-                await this.PublishPolicyUpdate(update);
+                await this.ConfigUpdateHandler(config);
             });
         }
 
         async Task ConfigUpdateHandler(EdgeHubConfig config)
         {
-            PolicyUpdate update = ConfigToPolicyUpdate(config);
-            await this.PublishPolicyUpdate(update);
+            try
+            {
+                PolicyUpdate update = ConfigToPolicyUpdate(config);
+
+                Events.PublishPolicyUpdate(update);
+
+                var payload = Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(update));
+                await this.connector.SendAsync(Topic, payload);
+            }
+            catch (Exception ex)
+            {
+                Events.ErrorUpdatingPolicy(ex);
+            }
         }
 
         static PolicyUpdate ConfigToPolicyUpdate(EdgeHubConfig config)
         {
-            Option<AuthorizationConfig> maybePolicy = config.BrokerConfiguration.AndThen(
-                config => config.Authorizations.Map(policy => policy));
+            Option<AuthorizationConfig> maybePolicy = config.BrokerConfiguration.FlatMap(
+                config => config.Authorizations);
 
             return maybePolicy.Match(
                 policy => new PolicyUpdate(JsonConvert.SerializeObject(policy)),
@@ -71,20 +81,6 @@ namespace Microsoft.Azure.Devices.Edge.Hub.MqttBrokerAdapter
                 'schemaVersion': '2020-10-30',
                 'statements': [ ]
             }");
-        }
-
-        async Task PublishPolicyUpdate(PolicyUpdate update)
-        {
-            Events.PublishPolicyUpdate(update);
-            try
-            {
-                var payload = Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(update));
-                await this.connector.SendAsync(Topic, payload);
-            }
-            catch (Exception ex)
-            {
-                Events.ErrorUpdatingPolicy(ex);
-            }
         }
 
         static class Events
@@ -107,6 +103,25 @@ namespace Microsoft.Azure.Devices.Edge.Hub.MqttBrokerAdapter
             {
                 Log.LogError((int)EventIds.ErrorUpdatingPolicy, ex, $"A problem occurred while updating authorization policy to mqtt broker.");
             }
+        }
+
+        /// <summary>
+        /// PolicyUpdate is a Data Transfer Object used for sending authorization policy
+        /// definition from EdgeHub core to Mqtt Broker.
+        /// </summary>
+        internal class PolicyUpdate
+        {
+            [JsonConstructor]
+            public PolicyUpdate(string policy)
+            {
+                this.Definition = Preconditions.CheckNonWhiteSpace(policy, nameof(policy));
+            }
+
+            /// <summary>
+            /// A string that contains new policy definition in json format.
+            /// </summary>
+            [JsonProperty]
+            public string Definition { get; }
         }
     }
 }
